// client/src/app/models/SummaryModels.ts
export class GenderSummary{
	femaleCount: number;
	maleCount: number;
}

export class TermAvgModel{
	t1Avg: number;
	t2Avg: number;
}

export class ExtraActivitySummary{
	extraYes: number;
	extraNo: number;
}

export class AgeToWorktimeSummary{
	workTime: number;
	age: number;
}

export class AgeToInternetSummary{
	internet: number;
	age: number;
}

export class AgeToFreetimeSummary{
	freetime: number;
	age: number;
}


export class AgeSummary{
	ageCount: number;
	age: number;
}


export class SumSummary{
	sum: number;
	avg: number;
}